move to right place
